use a separate export for the edge-light environment
