+++ packages/render/src/react-internals.d.ts
@@ -1,3 +1,6 @@
 declare module "react-dom/server.browser" {
   export * from "react-dom/server";
 }
+declare module "react-dom/server.edge" {
+  export * from "react-dom/server";
+}
