+++ packages/render/tsup.config.ts
@@ -9,6 +9,12 @@ export default defineConfig([
   },
   {
     dts: true,
+    entry: ["./src/edge-light/index.ts"],
+    outDir: "./dist/edge-light",
+    format: ["cjs", "esm"],
+  },
+  {
+    dts: true,
     entry: ["./src/browser/index.ts"],
     outDir: "./dist/browser",
     format: ["cjs", "esm"],
